+++ src/main/java/ee/meriloo/toru/model/Pipeline.java
@@ -1,18 +1,19 @@
 package ee.meriloo.toru.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class Pipeline {
 
-    private List<Double> pipes;
+    private List<BigDecimal> pipes;
     private List<Junction> junctions;
 
-    public Pipeline(List<Double> pipes, List<Junction> junctions) {
+    public Pipeline(List<BigDecimal> pipes, List<Junction> junctions) {
         this.pipes = pipes;
         this.junctions = junctions;
     }
 
-    public List<Double> getPipes() {
+    public List<BigDecimal> getPipes() {
         return pipes;
     }
